Refactor: 게시글 작성 컴포넌트에서 데이터 생성 로직을 훅으로 분리 #41
